backend/requirements.txt
aioredis==2.0.0
anyio==3.6.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
black==22.10.0
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
cleo==1.0.0a5
click==8.1.3
colorama==0.4.5
crashtest==0.3.1
Deprecated==1.2.13
exceptiongroup==1.0.0rc9
fastapi
fastapi-cache2==0.1.9
fastapi-jwt-auth==0.5.0
google-auth==2.11.0
gunicorn==20.1.0
h11==0.12.0
hiredis
httpcore==0.15.0
httpx==0.23.0
idna==3.4
iniconfig==1.1.1
more-itertools==8.14.0
mypy-extensions==0.4.3
outcome==1.2.0
packaging==21.3
pathspec==0.10.1
pendulum==2.1.2
platformdirs==2.5.2
pluggy==1.0.0
pptree==3.1
psycopg2-binary
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
PyJWT==1.7.1
pylev==1.4.0
pyparsing==3.0.9
pytest==7.1.3
pytest-trio==0.7.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
python-ulid==1.1.0
pytzdata==2020.1
redis
redis-om
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a27
sqlmodel==0.0.8
starlette==0.20.4
tomli==2.0.1
trio==0.22.0
types-redis==4.3.21.3
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn
wrapt==1.14.1
